# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Mist Automation & Backup API
  description: |
    Comprehensive web application for automating Juniper Mist operations and managing configuration backups.
    
    ## Authentication
    This API uses JWT Bearer token authentication. To authenticate:
    1. Call `POST /api/v1/auth/login` with your credentials
    2. Use the returned `access_token` in the `Authorization` header as `Bearer <token>`
    
    ## First-Time Setup (Onboarding)
    When the system has no users, use `POST /api/v1/auth/onboard` to create the first admin user.
    
    ## Rate Limiting
    API requests are rate-limited per user. Please handle 429 responses appropriately.
    Login attempts are rate-limited to 5 attempts per 5 minutes per IP/email combination.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: API v1
...
security:
  - BearerAuth: []
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: List users
      description: List all users (admin only)
      operationId: listUsers
      parameters:
        - name: skip
          in: query
          description: Number of users to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Number of users to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: is_active
          in: query
          description: Filter by active status
          schema:
            type: boolean
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Users
      summary: Create user
      description: Create a new user (admin only)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid password or roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{user_id}:
    get:
      tags:
        - Users
      summary: Get user
      description: Get user details by ID (admin only)
      operationId: getUser
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid user ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Users
      summary: Update user
      description: Update user details (admin only)
      operationId: updateUser
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid user ID format or cannot disable own account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Users
      summary: Delete user
      description: Delete a user (admin only)
      operationId: deleteUser
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      responses:
        '204':
          description: User deleted
        '400':
          description: Invalid user ID format or cannot delete self
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ═══════════════════════════════════════════════════════════════════════════
  # Workflow Endpoints
  # ═══════════════════════════════════════════════════════════════════════════
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
...
    UserCreate:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          format: password
          minLength: 1
          description: User password
        roles:
          type: array
          items:
            type: string
            enum: [admin, automation, backup]
          default: []
          description: "User roles. Allowed values: admin, automation, backup"
        timezone:
          type: string
          default: UTC
          description: User timezone
    UserUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User email address
        roles:
          type: array
          items:
            type: string
            enum: [admin, automation, backup]
          description: "User roles. Allowed values: admin, automation, backup"
        timezone:
          type: string
          description: User timezone
        is_active:
          type: boolean
          description: Whether user is active
    UserResponse:
      type: object
      required:
        - id
        - email
        - roles
        - timezone
        - is_active
        - totp_enabled
        - created_at
      properties:
        id:
          type: string
          description: User ID
        email:
          type: string
          format: email
          description: User email
        roles:
          type: array
          items:
            type: string
          description: User roles
        timezone:
          type: string
          description: User timezone
        is_active:
          type: boolean
          description: Whether user is active
        totp_enabled:
          type: boolean
          description: Whether 2FA is enabled
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        last_login:
          type:
            - string
            - 'null'
          format: date-time
          description: Last login timestamp
    UserListResponse:
      type: object
      required:
        - users
        - total
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
          description: List of users
        total:
          type: integer
          description: Total number of users

    # ─────────────────────────────────────────────────────────────────────────
    # Workflow
    # ─────────────────────────────────────────────────────────────────────────
...